+++ lib/agens/serving.ex
@@ -16,39 +16,41 @@ defmodule Agens.Serving do
     The Config struct represents the configuration for a Serving process.
 
     ## Fields
-    - `:name` - The name of the `Agens.Serving` process.
+    - `:name` - The unique name for the Serving process.
     - `:serving` - The `Nx.Serving` struct or `GenServer` module for the `Agens.Serving`.
-    - `:prompts` - A map of custom prompt prefixes. If `nil`, default prompt prefixes will be used instead. Default prompt prefixes can also be overridden by using the `prompts` options in `Agens.Supervisor`.
+    - `:prefixes` - An `Agens.Prefixes` struct of custom prompt prefixes. If `nil`, default prompt prefixes will be used instead. Default prompt prefixes can also be overridden by using the `prefixes` options in `Agens.Supervisor`.
+    - `:finalize` - A function that accepts the prepared prompt (including any applied prefixes) and returns a modified version of the prompt. Useful for wrapping the prompt or applying final processing before sending to the LM for inference. If `nil`, the prepared prompt will be used as-is.
+    - `:args` - Additional arguments to be passed to the `Nx.Serving` or `GenServer` module. See the [Nx.Serving](https://hexdocs.pm/nx/Nx.Serving.html) or [GenServer](https://hexdocs.pm/elixir/GenServer.html) documentation for more information.
     """
 
     @type t :: %__MODULE__{
             name: atom(),
             serving: Nx.Serving.t() | module(),
-            prompts: map() | nil
+            args: keyword(),
+            prefixes: Agens.Prefixes.t() | nil,
+            finalize: (String.t() -> String.t()) | nil
           }
 
     @enforce_keys [:name, :serving]
-    defstruct [:name, :serving, :prompts]
+    defstruct [:name, :serving, :prefixes, :finalize, args: []]
   end
 
   defmodule State do
     @moduledoc false
 
     @type t :: %__MODULE__{
-            registry: atom(),
             config: Config.t()
           }
 
-    @enforce_keys [:registry, :config]
-    defstruct [:registry, :config]
+    @enforce_keys [:config]
+    defstruct [:config]
   end
 
   use GenServer
 
   alias Agens.Message
 
-  @suffix "Supervisor"
-  @parent "Wrapper"
+  @suffix "Serving"
 
   # ===========================================================================
   # Public API
@@ -68,9 +70,7 @@ defmodule Agens.Serving do
   @spec stop(atom()) :: :ok | {:error, :serving_not_found}
   def stop(name) when is_atom(name) do
     name
-    |> parent_name()
     |> Agens.name_to_pid({:error, :serving_not_found}, fn pid ->
-      GenServer.call(pid, {:stop, name})
       :ok = DynamicSupervisor.terminate_child(Agens, pid)
     end)
   end
@@ -81,7 +81,6 @@ defmodule Agens.Serving do
   @spec get_config(atom() | pid()) :: {:ok, Config.t()} | {:error, :serving_not_found}
   def get_config(name) when is_atom(name) do
     name
-    |> parent_name()
     |> Agens.name_to_pid({:error, :serving_not_found}, fn pid -> get_config(pid) end)
   end
 
@@ -95,10 +94,22 @@ defmodule Agens.Serving do
   @spec run(Message.t()) :: String.t() | {:error, :serving_not_found}
   def run(%Message{serving_name: name} = message) when is_atom(name) do
     name
-    |> parent_name()
     |> Agens.name_to_pid({:error, :serving_not_found}, fn pid ->
       GenServer.call(pid, {:run, message})
     end)
+  end
+
+  @doc false
+  @spec finalize(atom() | pid(), String.t()) :: {:ok, String.t()} | {:error, :serving_not_found}
+  def finalize(name, prompt) when is_atom(name) do
+    name
+    |> Agens.name_to_pid({:error, :serving_not_found}, fn pid ->
+      finalize(pid, prompt)
+    end)
+  end
+
+  def finalize(pid, prompt) when is_pid(pid) do
+    {:ok, GenServer.call(pid, {:finalize, prompt})}
   end
 
   # ===========================================================================
@@ -108,42 +119,32 @@ defmodule Agens.Serving do
   @doc false
   @spec child_spec(Config.t()) :: Supervisor.child_spec()
   def child_spec(%Config{} = config) do
-    name = parent_name(config.name)
-
     %{
-      id: name,
-      start: {__MODULE__, :start_link, [config]},
-      type: :worker,
-      restart: :transient
+      id: config.name,
+      start: {__MODULE__, :start_link, [config]}
     }
   end
 
   @doc false
   @spec start_link(keyword(), Config.t()) :: GenServer.on_start()
   def start_link(extra, config) do
-    name = parent_name(config.name)
     opts = Keyword.put(extra, :config, config)
-    GenServer.start_link(__MODULE__, opts, name: name)
+    GenServer.start_link(__MODULE__, opts, name: config.name)
   end
 
   @doc false
   @impl true
   @spec init(keyword()) :: {:ok, State.t()} | {:stop, term(), State.t()}
   def init(opts) do
-    registry = Keyword.fetch!(opts, :registry)
-    prompts = Keyword.fetch!(opts, :prompts)
+    prefixes = Keyword.fetch!(opts, :prefixes)
     config = Keyword.fetch!(opts, :config)
-    config = if is_nil(config.prompts), do: Map.put(config, :prompts, prompts), else: config
-    state = %State{config: config, registry: registry}
-    {m, f, a} = start_function(config)
+    config = if is_nil(config.prefixes), do: Map.put(config, :prefixes, prefixes), else: config
+    state = %State{config: config}
 
-    m
-    |> apply(f, a)
+    config
+    |> start_serving()
     |> case do
       {:ok, pid} when is_pid(pid) ->
-        name = serving_name(config.name)
-        {:ok, _} = Registry.register(registry, name, {pid, config})
-
         {:ok, state}
 
       {:error, reason} ->
@@ -154,15 +155,6 @@ defmodule Agens.Serving do
   # ===========================================================================
   # Callbacks
   # ===========================================================================
-
-  @doc false
-  @impl true
-  @spec handle_call({:stop, atom()}, {pid, term}, State.t()) :: {:reply, :ok, State.t()}
-  def handle_call({:stop, serving_name}, _from, state) do
-    serving_name = serving_name(serving_name)
-    Registry.unregister(state.registry, serving_name)
-    {:reply, :ok, state}
-  end
 
   @doc false
   @impl true
@@ -180,6 +172,20 @@ defmodule Agens.Serving do
     {:reply, result, state}
   end
 
+  @doc false
+  @impl true
+  @spec handle_call({:finalize, String.t()}, {pid, term}, State.t()) ::
+          {:reply, String.t(), State.t()}
+  def handle_call({:finalize, prompt}, _, %State{config: %Config{finalize: finalize}} = state) do
+    final =
+      case finalize do
+        fun when is_function(fun, 1) -> fun.(prompt)
+        _ -> prompt
+      end
+
+    {:reply, final, state}
+  end
+
   # ===========================================================================
   # Private
   # ===========================================================================
@@ -188,6 +194,7 @@ defmodule Agens.Serving do
   @spec do_run(Config.t(), Message.t()) :: String.t()
   defp do_run(%Config{serving: %Nx.Serving{}}, %Message{} = message) do
     message.serving_name
+    |> serving_name()
     |> Nx.Serving.batched_run(message.prompt)
     |> case do
       %{results: [%{text: result}]} -> result
@@ -196,30 +203,32 @@ defmodule Agens.Serving do
   end
 
   defp do_run(_, %Message{} = message) do
-    serving_name = serving_name(message.serving_name)
-    # need to get pid?
-    GenServer.call(serving_name, {:run, message})
+    message.serving_name
+    |> serving_name()
+    |> GenServer.call({:run, message})
   end
 
   @doc false
-  @spec start_function(Config.t()) :: tuple()
-  defp start_function(%Config{serving: %Nx.Serving{} = serving} = config) do
-    {Nx.Serving, :start_link, [[serving: serving, name: config.name]]}
-  end
-
-  @doc false
-  @spec start_function(Config.t()) :: tuple()
-  # Module.concat with "Supervisor" for Nx.Serving parity
-  defp start_function(%Config{serving: serving} = config) when is_atom(serving) do
+  @spec start_serving(Config.t()) :: tuple()
+  defp start_serving(%Config{serving: %Nx.Serving{} = serving, args: args} = config) do
     name = serving_name(config.name)
-    {serving, :start_link, [[name: name, config: config]]}
+
+    opts =
+      args
+      |> Keyword.put(:serving, serving)
+      |> Keyword.put(:name, name)
+
+    Nx.Serving.start_link(opts)
+  end
+
+  defp start_serving(%Config{serving: serving, args: args} = config) when is_atom(serving) do
+    name = serving_name(config.name)
+    opts = Keyword.put(args, :name, name)
+
+    GenServer.start_link(serving, config, opts)
   end
 
   @doc false
   @spec serving_name(atom) :: atom
   defp serving_name(name) when is_atom(name), do: Module.concat(name, @suffix)
-
-  @doc false
-  @spec parent_name(atom) :: atom
-  defp parent_name(name) when is_atom(name), do: Module.concat(name, @parent)
 end