+++ test/agens/serving_test.exs
@@ -4,8 +4,7 @@ defmodule Agens.ServingTest do
   alias Agens.{Message, Serving}
 
   defp start_agens(_ctx) do
-    opts = [registry: Agens.CustomRegistry]
-    {:ok, _pid} = start_supervised({Agens.Supervisor, name: Agens.Supervisor, opts: opts})
+    {:ok, _pid} = start_supervised({Agens.Supervisor, name: Agens.Supervisor})
     :ok
   end
 
@@ -102,7 +101,7 @@ defmodule Agens.ServingTest do
 
       batch = Nx.Batch.stack([Nx.tensor([1, 2, 3])])
 
-      message = %Message{serving_name: serving_name, prompt: batch}
+      message = %Message{serving_name: serving_name, prompt: batch, input: "not used"}
 
       assert %Nx.Tensor{
                type: {:s, 64},